fix: app-module first-compile errors
- AppContainer: PeriodicWorkRequestBuilder needs Long duration (hours.toLong())
- MonitoringScreen: missing Box import
- SettingsScreen: restructure token dialog inside the composable function
  (brace nesting left it at top level), add AlertDialog/TextButton imports
- HashingTest: wrap suspend sha256 calls in runBlocking

+++ app/src/main/kotlin/com/localai/runtime/runtime/AppContainer.kt
@@ -242,7 +242,7 @@ class AppContainer(val context: Context) {
 
     fun scheduleCatalogRefresh(hours: Int) {
         val request = PeriodicWorkRequestBuilder<CatalogRefreshWorker>(
-            hours.coerceAtLeast(1), TimeUnit.HOURS,
+            hours.coerceAtLeast(1).toLong(), TimeUnit.HOURS,
         )
             .setConstraints(
                 Constraints.Builder()
